backend/requirements.txt
# Flask and extensions
Flask==3.0.0
flask-cors==4.0.0

# Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# AI Integration
openai==1.6.1
google-generativeai==0.3.2
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Utilities
Werkzeug==3.0.1
